Refactor Propsを設定し直した
